feat(editor): shared block editor + template/registry de-hardcoding (#529)
* fix(templates): require productName/actionUrl/supportEmail for onboarding-welcome

The onboarding-welcome React Email template silently fell back to
"Opensend" for productName, "https://opensend.dev/docs" for actionUrl,
and "support@example.com" for supportEmail when senders didn't pass
those variables — putting the platform's brand and a placeholder
support address into customer emails.

Make all three required at the registry level and validate them in
renderReactEmailTemplate so missing inputs raise a typed
react_required_variable_missing error instead of producing a misleading
message. The demo-welcome fixture keeps its optional defaults but now
uses neutral placeholders ("your product", "#") rather than
Opensend-branded URLs.


* refactor(editor): extract shared block editor canvas

Pull the Notion-style block model out of broadcast-editor.tsx into a
reusable src/components/block-editor module so the template editor can
mount the same visual canvas. Exports BlockEditorCanvas, BlockRenderer,
blocksToHtml serializer, and the block type catalog. broadcast-editor
now imports the shared component; behavior preserved (26 tests pass).

* feat(templates): mount shared block editor in /templates editor

Wire the BlockEditorCanvas into the template editor with a Visual/Code
tab toggle. Visual mode renders the same canvas the broadcast editor
uses (slash menu, components, variables, toolbar tabs); Code mode keeps
the HTML textarea for raw edits. Blocks serialize to HTML via
blocksToHtml on every change so the existing PATCH body shape is
preserved. Test switches to Code tab before editing the HTML textarea.

---------
