+++ src/components/block-editor/index.ts
@@ -1,0 +1,12 @@
+export { BlockEditorCanvas } from "./block-editor-canvas";
+export { BlockRenderer } from "./block-renderer";
+export { blocksToHtml } from "./serialize";
+export {
+  COMPONENT_ITEMS,
+  SLASH_MENU_ITEMS,
+  VARIABLES,
+  type BlockType,
+  type EditorBlock,
+  type SlashMenuCategory,
+  type SlashMenuItem,
+} from "./types";
